+++ ExampleMicroService/Commands/HaalPolissenOpCommand.cs
@@ -10,13 +10,6 @@ namespace ExampleMicroService.Commands
     public class HaalPolissenOpCommand : DomainCommand
     {
         /// <summary>
-        /// A list of Polisses that is initially empty and we intend
-        /// to fill in the destination service that will receive this
-        /// command.
-        /// </summary>
-        public IEnumerable<Polis> Polisses = new List<Polis>();
-
-        /// <summary>
         /// To ensure all HaalPolissenOpCommands have the same destination queue,
         /// we pass the name of the queue in the constructor.
         ///