+++ Minor.Miffy.MicroServices.Test.Conventions/Component/Commands/GetAnimalsCommand.cs
@@ -1,0 +1,11 @@
+using Minor.Miffy.MicroServices.Commands;
+
+namespace Minor.Miffy.MicroServices.Test.Conventions.Component.Commands
+{
+    public class GetAnimalsCommand : DomainCommand
+    {
+        public GetAnimalsCommand() : base("command.test")
+        {
+        }
+    }
+}