+++ Minor.Miffy.MicroServices.Test.Conventions/Component/Models/Animal.cs
@@ -1,0 +1,8 @@
+namespace Minor.Miffy.MicroServices.Test.Conventions.Component.Models
+{
+    public class Animal
+    {
+        public string Species { get; set; }
+        public string Name { get; set; }
+    }
+}