+++ Minor.Miffy.MicroServices/Commands/CommandPublisher.cs
@@ -14,7 +14,7 @@ namespace Minor.Miffy.MicroServices.Commands
         /// <summary>
         /// Sender to send a message through the bus
         /// </summary>
-        private readonly ICommandSender _sender;
+        private readonly Miffy.ICommandSender _sender;
 
         /// <summary>
         /// Logger
@@ -32,9 +32,18 @@ namespace Minor.Miffy.MicroServices.Commands
         }
 
         /// <summary>
-        /// Publish a domain event
+        /// Synchronous variant of the PublishAsync that basically only
+        /// returns the result of the async one.
         /// </summary>
-        public async Task<T> PublishAsync<T>(DomainCommand domainCommand)
+        public TReturn Publish<TReturn>(DomainCommand domainCommand)
+        {
+            return PublishAsync<TReturn>(domainCommand).Result;
+        }
+
+        /// <summary>
+        /// Publish a domain command with a specific return result
+        /// </summary>
+        public async Task<TReturn> PublishAsync<TReturn>(DomainCommand domainCommand)
         {
             _logger.LogTrace(
                 $"Publishing domain command with type {domainCommand.GetType().Name} and ID {domainCommand.Id}");
@@ -56,8 +65,8 @@ namespace Minor.Miffy.MicroServices.Commands
 
             try
             {
-                string jsonBody = Encoding.Unicode.GetString(result.Body);
-                return (T) JsonConvert.DeserializeObject(jsonBody, typeof(T));
+                string jsonBody = Encoding.Unicode.GetString(result?.Body);
+                return (TReturn) JsonConvert.DeserializeObject(jsonBody, typeof(TReturn));
             }
             catch (ArgumentNullException exception)
             {