+++ Minor.Miffy.MicroServices/Commands/ICommandPublisher.cs
@@ -5,5 +5,6 @@ namespace Minor.Miffy.MicroServices.Commands
     public interface ICommandPublisher
     {
         Task<T> PublishAsync<T>(DomainCommand domainCommand);
+        T Publish<T>(DomainCommand domainCommand);
     }
 }